+++ src/Sensorario/ValueObject/Validators/RightType.php
@@ -31,8 +31,11 @@ final class RightType implements Validator
                         . '`'
                     );
                 }
-
-                if (get_class($valueObject->get($key)) != current($type)) {
+                
+                if (
+                    !($valueObject->get($key) instanceof \Sensorario\ValueObject\ValueObject) &&
+                    get_class($valueObject->get($key)) != current($type)
+                )  {
                     throw new RuntimeException(
                         'Attribute `' . $key
                         . '` must be an object of type ' . current($type)